Revert "lane detection using async api"
